ntic2_source/ntic2_ai_agent_production/backend/requirements.txt

flask
psycopg2-binary
openai
langchain
chromadb
python-dotenv
beautifulsoup4
requests
PyPDF2
pdfplumber
openpyxl
pandas
Pillow
pytesseract
sentence-transformers
